+++ src/prelude/traits.rs
@@ -45,7 +45,9 @@ pub use crate::traits::applicative::Applicative;
 pub use crate::traits::bifunctor::Bifunctor;
 pub use crate::traits::foldable::Foldable;
 pub use crate::traits::functor::Functor;
-pub use crate::traits::hkt::{BinaryHKT, HKT};
+#[allow(deprecated)]
+pub use crate::traits::hkt::BinaryHKT;
+pub use crate::traits::hkt::HKT;
 #[allow(deprecated)]
 pub use crate::traits::iso::Iso;
 pub use crate::traits::monad::Monad;